3.6.13 abstract output
